+++ src/components/mosaic/Grid/Grid.jsx
@@ -16,18 +16,21 @@ import { Row } from '../../../components';
  * @param {Object[]} props.rows Rows in the grid.
  * @param {func} props.selectTile Select tile method.
  * @param {func} props.setHovered Set hovered tile method.
+ * @param {func} props.handleDrop Handle drop tile event.
  * @param {func} props.setTileContent Set tile content method.
  * @returns {string} Markup of the row.
  */
-const Grid = ({ rows, selectTile, setHovered, setTileContent }) => (
-  <UIGrid>
+const Grid = ({ rows, selectTile, setHovered, handleDrop, setTileContent }) => (
+  <UIGrid className="mosaic">
     {rows.map((row, index) => (
       <Row
         row={index}
         selectTile={selectTile}
         setHovered={setHovered}
+        handleDrop={handleDrop}
         setTileContent={setTileContent}
         columns={row.columns}
+        hovered={row.hovered}
       />
     ))}
   </UIGrid>
@@ -43,6 +46,7 @@ Grid.propTypes = {
     .isRequired,
   selectTile: PropTypes.func.isRequired,
   setHovered: PropTypes.func.isRequired,
+  handleDrop: PropTypes.func.isRequired,
   setTileContent: PropTypes.func.isRequired,
 };