+++ src/components/mosaic/Grid/Grid.test.jsx
@@ -10,6 +10,7 @@ test('renders a grid component', () => {
       rows={[{ colums: [] }]}
       selectTile={() => {}}
       setHovered={() => {}}
+      handleDrop={() => {}}
       setTileContent={() => {}}
     />,
   );