+++ src/components/mosaic/Row/Row.jsx
@@ -15,21 +15,33 @@ import { Column } from '../../../components';
  * @param {Object} props Component properties.
  * @param {Object} props.columns Column data.
  * @param {number} props.row Row index.
+ * @param {string} props.hovered Hovered state.
  * @param {func} props.selectTile Select tile method.
  * @param {func} props.setHoverd Set hovered tile method.
+ * @param {func} props.handleDrop Handle tile drop event.
  * @param {func} props.setTileContent Set tile content method.
  * @returns {string} Markup of the row.
  */
-const Row = ({ columns, row, selectTile, setHovered, setTileContent }) => (
-  <Grid.Row>
+const Row = ({
+  columns,
+  row,
+  hovered,
+  selectTile,
+  setHovered,
+  handleDrop,
+  setTileContent,
+}) => (
+  <Grid.Row className={hovered || ''}>
     {columns.map((column, index) => (
       <Column
         row={row}
         column={index}
         width={column.width}
         tiles={column.tiles}
+        hovered={column.hovered}
         selectTile={selectTile}
         setHovered={setHovered}
+        handleDrop={handleDrop}
         setTileContent={setTileContent}
       />
     ))}
@@ -44,9 +56,20 @@ const Row = ({ columns, row, selectTile, setHovered, setTileContent }) => (
 Row.propTypes = {
   columns: PropTypes.arrayOf(PropTypes.object).isRequired,
   row: PropTypes.number.isRequired,
+  hovered: PropTypes.string,
   selectTile: PropTypes.func.isRequired,
   setHovered: PropTypes.func.isRequired,
+  handleDrop: PropTypes.func.isRequired,
   setTileContent: PropTypes.func.isRequired,
+};
+
+/**
+ * Default props.
+ * @property {Object} defaultProps Default properties.
+ * @static
+ */
+Row.defaultProps = {
+  hovered: null,
 };
 
 export default Row;